+++ src/anyio/abc/_sockets.py
@@ -10,6 +10,7 @@ from ipaddress import IPv4Address, IPv6Address
 from socket import AddressFamily
 from typing import Any, TypeVar, Union
 
+from .._core._eventloop import get_async_backend
 from .._core._typedattr import (
     TypedAttributeProvider,
     TypedAttributeSet,
@@ -29,6 +30,51 @@ SockAddrType: TypeAlias = Union[IPSockAddrType, str]
 UDPPacketType: TypeAlias = tuple[bytes, IPSockAddrType]
 UNIXDatagramPacketType: TypeAlias = tuple[bytes, str]
 T_Retval = TypeVar("T_Retval")
+
+
+def _validate_socket(
+    sock_or_fd: socket.socket | int,
+    sock_type: socket.SocketKind,
+    addr_family: socket.AddressFamily = socket.AF_UNSPEC,
+    *,
+    require_connected: bool = False,
+    require_bound: bool = False,
+) -> socket.socket:
+    if isinstance(sock_or_fd, int):
+        sock = socket.socket(fileno=sock_or_fd)
+    elif isinstance(sock_or_fd, socket.socket):
+        sock = sock_or_fd
+    else:
+        raise TypeError(
+            f"expected an int or socket, got {type(sock_or_fd).__qualname__} instead"
+        )
+
+    if addr_family != socket.AF_UNSPEC and sock.family != addr_family:
+        raise ValueError(
+            f"address family mismatch: expected {addr_family.name}, got "
+            f"{sock.family.name}"
+        )
+
+    if sock.type != sock_type:
+        raise ValueError(
+            f"socket type mismatch: expected {sock_type.name}, got {sock.type.name}"
+        )
+
+    if require_connected:
+        try:
+            sock.getpeername()
+        except OSError:
+            raise ValueError("the socket must be connected") from None
+
+    if require_bound:
+        if (sock.family == socket.AF_UNIX and not sock.getsockname()) or (
+            sock.family in (socket.AF_INET, socket.AF_INET6)
+            and sock.getsockname()[1] == 0
+        ):
+            raise ValueError("the socket must be bound to a local address")
+
+    sock.setblocking(False)
+    return sock
 
 
 class SocketAttribute(TypedAttributeSet):
@@ -91,8 +137,40 @@ class SocketStream(ByteStream, _SocketProvider):
     Supports all relevant extra attributes from :class:`~SocketAttribute`.
     """
 
+    @classmethod
+    async def from_socket(cls, sock_or_fd: socket.socket | int) -> SocketStream:
+        """
+        Wrap an existing socket object or file descriptor as a socket stream.
+
+        The newly created socket wrapper takes ownership of the socket being passed in.
+        The existing socket must already be connected.
+
+        :param sock_or_fd: a socket object or file descriptor
+        :return: a socket stream
+
+        """
+        sock = _validate_socket(sock_or_fd, socket.SOCK_STREAM, require_connected=True)
+        return await get_async_backend().wrap_stream_socket(sock)
+
 
 class UNIXSocketStream(SocketStream):
+    @classmethod
+    async def from_socket(cls, sock_or_fd: socket.socket | int) -> UNIXSocketStream:
+        """
+        Wrap an existing socket object or file descriptor as a UNIX socket stream.
+
+        The newly created socket wrapper takes ownership of the socket being passed in.
+        The existing socket must already be connected.
+
+        :param sock_or_fd: a socket object or file descriptor
+        :return: a UNIX socket stream
+
+        """
+        sock = _validate_socket(
+            sock_or_fd, socket.SOCK_STREAM, socket.AF_UNIX, require_connected=True
+        )
+        return await get_async_backend().wrap_unix_stream_socket(sock)
+
     @abstractmethod
     async def send_fds(self, message: bytes, fds: Collection[int | IOBase]) -> None:
         """
@@ -121,6 +199,23 @@ class SocketListener(Listener[SocketStream], _SocketProvider):
     Supports all relevant extra attributes from :class:`~SocketAttribute`.
     """
 
+    @classmethod
+    async def from_socket(
+        cls,
+        sock_or_fd: socket.socket | int,
+    ) -> SocketListener:
+        """
+        Wrap an existing socket object or file descriptor as a socket listener.
+
+        The newly created listener takes ownership of the socket being passed in.
+
+        :param sock_or_fd: a socket object or file descriptor
+        :return: a socket listener
+
+        """
+        sock = _validate_socket(sock_or_fd, socket.SOCK_STREAM, require_bound=True)
+        return await get_async_backend().wrap_listener_socket(sock)
+
     @abstractmethod
     async def accept(self) -> SocketStream:
         """Accept an incoming connection."""
@@ -148,6 +243,20 @@ class UDPSocket(UnreliableObjectStream[UDPPacketType], _SocketProvider):
     Supports all relevant extra attributes from :class:`~SocketAttribute`.
     """
 
+    @classmethod
+    async def from_socket(cls, sock_or_fd: socket.socket | int) -> UDPSocket:
+        """
+        Wrap an existing socket object or file descriptor as a UDP socket.
+
+        The newly created socket wrapper takes ownership of the socket being passed in.
+
+        :param sock_or_fd: a socket object or file descriptor
+        :return: a UDP socket
+
+        """
+        sock = _validate_socket(sock_or_fd, socket.SOCK_DGRAM)
+        return await get_async_backend().wrap_udp_socket(sock)
+
     async def sendto(self, data: bytes, host: str, port: int) -> None:
         """
         Alias for :meth:`~.UnreliableObjectSendStream.send` ((data, (host, port))).
@@ -163,6 +272,25 @@ class ConnectedUDPSocket(UnreliableObjectStream[bytes], _SocketProvider):
     Supports all relevant extra attributes from :class:`~SocketAttribute`.
     """
 
+    @classmethod
+    async def from_socket(cls, sock_or_fd: socket.socket | int) -> ConnectedUDPSocket:
+        """
+        Wrap an existing socket object or file descriptor as a connected UDP socket.
+
+        The newly created socket wrapper takes ownership of the socket being passed in.
+        The existing socket must already be connected.
+
+        :param sock_or_fd: a socket object or file descriptor
+        :return: a connected UDP socket
+
+        """
+        sock = _validate_socket(
+            sock_or_fd,
+            socket.SOCK_DGRAM,
+            require_connected=True,
+        )
+        return await get_async_backend().wrap_connected_udp_socket(sock)
+
 
 class UNIXDatagramSocket(
     UnreliableObjectStream[UNIXDatagramPacketType], _SocketProvider
@@ -172,6 +300,24 @@ class UNIXDatagramSocket(
 
     Supports all relevant extra attributes from :class:`~SocketAttribute`.
     """
+
+    @classmethod
+    async def from_socket(
+        cls,
+        sock_or_fd: socket.socket | int,
+    ) -> UNIXDatagramSocket:
+        """
+        Wrap an existing socket object or file descriptor as a UNIX datagram
+        socket.
+
+        The newly created socket wrapper takes ownership of the socket being passed in.
+
+        :param sock_or_fd: a socket object or file descriptor
+        :return: a UNIX datagram socket
+
+        """
+        sock = _validate_socket(sock_or_fd, socket.SOCK_DGRAM, socket.AF_UNIX)
+        return await get_async_backend().wrap_unix_datagram_socket(sock)
 
     async def sendto(self, data: bytes, path: str) -> None:
         """Alias for :meth:`~.UnreliableObjectSendStream.send` ((data, path))."""
@@ -184,3 +330,24 @@ class ConnectedUNIXDatagramSocket(UnreliableObjectStream[bytes], _SocketProvider
 
     Supports all relevant extra attributes from :class:`~SocketAttribute`.
     """
+
+    @classmethod
+    async def from_socket(
+        cls,
+        sock_or_fd: socket.socket | int,
+    ) -> ConnectedUNIXDatagramSocket:
+        """
+        Wrap an existing socket object or file descriptor as a connected UNIX datagram
+        socket.
+
+        The newly created socket wrapper takes ownership of the socket being passed in.
+        The existing socket must already be connected.
+
+        :param sock_or_fd: a socket object or file descriptor
+        :return: a connected UNIX datagram socket
+
+        """
+        sock = _validate_socket(
+            sock_or_fd, socket.SOCK_DGRAM, socket.AF_UNIX, require_connected=True
+        )
+        return await get_async_backend().wrap_connected_unix_datagram_socket(sock)